refactor(web_core): modular package entrypoints and version-agnostic markdown types
