+++ renderers/lit/src/0.8/core.ts
@@ -17,12 +17,11 @@
 export * as Events from './events/events.js';
 import * as Types from '@a2ui/web_core/types/types';
 import * as Guards from '@a2ui/web_core/data/guards';
-import {Schemas} from '@a2ui/web_core';
+import {Schemas, Events as WebEvents} from '@a2ui/web_core/v0_8';
 import * as Styles from '@a2ui/web_core/styles/index';
 import {A2uiMessageProcessor} from '@a2ui/web_core/data/model-processor';
 import * as Primitives from '@a2ui/web_core/types/primitives';
 import {create as createSignalA2uiMessageProcessor} from './data/signal-model-processor.js';
-import {Events as WebEvents} from '@a2ui/web_core';
 export {Types, Guards, Schemas, Styles, A2uiMessageProcessor, Primitives, WebEvents};
 
 export const Data = {
