+++ renderers/lit/src/v0_9/directives/markdown.ts
@@ -18,7 +18,7 @@ import {html, noChange} from 'lit';
 import {Directive, DirectiveParameters, Part, directive} from 'lit/directive.js';
 import {unsafeHTML} from 'lit/directives/unsafe-html.js';
 import {until} from 'lit/directives/until.js';
-import * as Types from '@a2ui/web_core/types/types';
+import type {MarkdownRenderer, MarkdownRendererOptions} from '@a2ui/web_core';
 
 class MarkdownDirective extends Directive {
   private lastValue: string | null = null;
@@ -45,8 +45,8 @@ class MarkdownDirective extends Directive {
    */
   render(
     value: string,
-    markdownRenderer?: Types.MarkdownRenderer,
-    markdownOptions?: Types.MarkdownRendererOptions,
+    markdownRenderer?: MarkdownRenderer,
+    markdownOptions?: MarkdownRendererOptions,
   ) {
     if (markdownRenderer) {
       const rendered = markdownRenderer(value, markdownOptions).then(value => {
