+++ typescript/web_core/src/expressions/index.ts
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      https://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,15 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import {createContext} from '@lit/context';
-import type {MarkdownRenderer} from '@a2ui/web_core';
-
-export type {MarkdownRenderer};
-
 /**
- * The markdown renderer context.
- *
- * This is used by the Text widget to render markdown content.
+ * Expression parser and evaluator for A2UI dynamic bindings.
  */
-export const markdown = createContext<MarkdownRenderer | undefined>(Symbol.for('A2UIMarkdown'));
+
+export * from './expression_parser.js';
