+++ typescript/web_core/src/v0_9/basic_catalog/context/markdown.ts
@@ -16,35 +16,13 @@
 
 import {createContext} from '@lit/context';
 
-/**
- * A map of tag names to a list of classnames to be applied to a tag.
- *
- * For example, if you want to apply the class "my-class" to all "h1" tags,
- * you would use `{"h1": ["my-class"]}`.
- */
-export type MarkdownRendererTagClassMap = Record<string, string[]>;
+import type {MarkdownRenderer} from '../../../rendering/markdown.js';
 
-/**
- * The options for the markdown renderer passed from A2UI.
- *
- * This includes the tagClassMap, which is a map of tag names to a list of
- * classnames to be applied to each tag, and configuration for the sanitizer.
- */
-export type MarkdownRendererOptions = {
-  tagClassMap?: MarkdownRendererTagClassMap;
-};
-
-/**
- * Renders `markdown` using `options`.
- *
- * Implementations MUST sanitize the resulting HTML to prevent XSS vulnerabilities.
- *
- * @returns A promise that resolves to the rendered, sanitized HTML as a string.
- */
-export type MarkdownRenderer = (
-  markdown: string,
-  options?: MarkdownRendererOptions,
-) => Promise<string>;
+export type {
+  MarkdownRenderer,
+  MarkdownRendererOptions,
+  MarkdownRendererTagClassMap,
+} from '../../../rendering/markdown.js';
 
 /**
  * The markdown renderer context.
